refactor: QueryBuilder 위치 변경
